Change when we add provisioner-provided repos.
Instead of adding them to the global list of repos to use, we instead
add them to each machine when we decide to install an OS on it.  This
allows provisioner repos to pe more easily provisioner-specific, and
eliminates having to add special-case code to handle skipping
provisioner-provided repos when they are not wanted.