Add delete and name change support to the provisioner data.
When a node is deleted, clear the data from the provisioner.

When a node changes name, rebuild the data by deleting the old
data and adding new data for the new name.

The provisioner is mostly working.  It leaves some files around, but
they are beign.  Will work to fix that as well.