Centralize event handling, and expose it via API.
This moves role callbacks from being handled via method dispatch with
subclass-overridable methods to a centralized event dispatching system,
and updates the in-core classes to produce and consume the appropriate
events.

Implementation-wise, the code is set up to handle registering event
sinks that will recieve events, and event selectors that handle
selecting events that a given sink is interested in.  API endpoints for
event sinks and selectors have been created at the expected locations.

For now, the only event sink we know how to handle is the in-process
endpoint (named 'inproc://'), and all roles in the core that used the
method-based event handling scheme have been updated to register the
appropriate selectors at barclamp import time.

By adding an appropriate external endpoint handler and nailing down
exactly what we expect a sink to handle in terms of RPCs or pub/sub, we
can do away with forcing roles that need custom behaviour to live in the
same process space as the core.

This also paves the way for adding auditing events and other useful things.