Add role preceeds list.
Preceeds indicates that when the role is bound to a node, it should
preceed (run before) roles in the preceeds list.  If the role it
preceeds has the implicit flag, then it will only preceed the noderole
that is bound to the node, otherwise it will preceed a randomly-chosen
noderole for the node it preceeds.